src/requirements.txt
# Dependencies
bson
dnspython
email_validator
flask
flask-login
flask-pymongo
flask-mongoengine
mock
mongoengine
pymongo
pyyaml
uWSGI
werkzeug
wtforms
